refactor: extract library module
