+++ CHANGELOG.md
@@ -8,4 +8,5 @@ All notable changes to this project will be documented in this file.
 - Updated documentation and tests.
 - Improved path validation and error handling in export script.
 - Export script now throws errors instead of exiting directly.
+- Added concurrency limit for loading files and split library code into `src/lib.ts`.
 - Added security note in README.
